requirements.txt
# ── Azure OpenAI ─────────────────────────────────────────────
openai>=1.51.0

# ── Vector Database ───────────────────────────────────────────
chromadb>=0.5.15

# ── FastAPI Backend ───────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.9.2

# ── Utilities ─────────────────────────────────────────────────
python-dotenv>=1.0.1
tiktoken>=0.8.0
colorama>=0.4.6
gunicorn>=23.0.0